food/requirements.txt
asgiref==3.5.0
Django==4.0.4
django-model-utils==4.2.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
